Remove restoremymaterials prop, add topo guessatom
In vis-charmm, I wrote a proc to alter the AOEdgy material properties.
This was giving me issues when not using VMD 1.9.2, so I just add a few
lines to explicitly edit the material properties.

Additionally, I added a command that guesses the atom types based on
mass, which makes it easier to do element-specific stuff, e.g., draw
hydrogen bonds and color by element.